Allow micawber to store provider schema/config in cache.
The various bootstrap_* functions may need to make an HTTP request to
bootstrap the provider config. This patch allows the provider config
response to be stored in the cache, potentially saving an HTTP request
at startup time. This can be disabled by passing refresh=True, which
will ignore cache and refetch the provider config.