Register endpoitns in reverse so that the most-specific is first.
Micawber uses an ordered dict of regex->provider. We have logic so that
the most-recently-added regex is selected when there are multiple
matches. Because oembed.com returns endpoints in most->least specific
order, we need to reverse the endpoints in order to ensure we choose the
most-specific regex first.

Fixes #101